requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.49
botocore==1.35.49
click==8.1.7
fastapi==0.115.3
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
